backend/pyproject.toml
[project]
name = "portfoliodb"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2",
    "django-cors-headers>=4.7.0",
    "djangorestframework>=3.16.0",
    "openbb>=4.1.3",
]

[tool.ruff]
# Ruff formatting settings
# Derived from tool.black chapter and .flake8 file
line-length = 120
exclude = [
    ".pytest_cache",
    "**/__pycache__/",
    "**/*.ipynb_checkpoints/",
    "*.ipynb",
]

[tool.ruff.lint]
extend-select = [
    "F",
    "I",
    "PL",
    "N",
] # enabling flake8, isort, pylint, pep8-naming
extend-ignore = [
    "D",
    "E501",
    "PLR0913",
    "E402",
    "C403",
    "F405",
    "E203",
    "E731",
    "E712",
]
